CEMS-2328 Implement an IncompleteEntityException
